工作历史表建立完成

+++ database/migrations/2019_05_31_135611_create_staffs_table.php
@@ -23,7 +23,7 @@ class CreateStaffsTable extends Migration
             $table->string('position_name')->nullable();
             $table->date('join_company')->nullable();
             $table->date('leave_company')->nullable();
-            $table->unsignedInteger('work_year')->nullable();
+            $table->decimal('work_year',5,2)->nullable();
             // $table->time('work_time');
             // $table->time('home_time');
             // $table->string('workdays',100); // Mon,Tue,Wed,Thu,Fri
